requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.20
pandas>=2.2.0
python-dotenv>=1.0.1
rich>=13.0.0
tenacity>=8.2.0
pydantic>=2.0.0
flake8>=7.0.0
pytest>=8.0.0
structlog>=24.1.0